+++ ssqueezepy/ssqueezing.py
@@ -13,7 +13,7 @@ def ssqueeze(Wx, w, ssq_freqs=None, scales=None, fs=None, t=None, transform='cwt
 
     # Arguments:
         Wx or Sx: np.ndarray
-            CWT or STFT of `x`. Wx is assumed L1-normed.
+            CWT or STFT of `x`. `Wx` is assumed L1-normed.
 
         w: np.ndarray
             Phase transform of `Wx` or `Sx`. Must be >=0.