+++ CHANGELOG.md
@@ -11,4 +11,4 @@
  - `cwt()` no longer returns `x_mean`
  - `padsignal` now only returns padded input by default; `get_params=True` for old behavior
  - Moved methods: `phase_cwt` & `phase_cwt_num` from `ssqueezing` to `_ssq_cwt`
- - _In future release_: return order of `cwt` and `stft` will be changed to have `Wx, dWx` and `Sx, dSx`
+ - _In future release_: return order of `cwt` and `stft` will be changed to have `Wx, dWx` and `Sx, dSx`, and `ssq_cwt` and `ssq_stft` to have `Tx, Wx` and `Tx, Sx`.